rename: 파일명에 component prefix를 제거한다 - components/applicant/Board - components/applicant/SortList - components/interview/Board - components/interview/SortList - components/common/SortList
