+++ frontend/components/applicant/Board.tsx
