+++ frontend/components/interview/Board.tsx
